fix compile dependency
Gradle scans all decorated plugin/tasks/extensions classes to find all the injection points, apply validations etc. This means that we cannot have any method signatures in the plugin code that references compile only dependencies. Simple fix is to move the affected code to a separate file. See: https://github.com/gradle/gradle/issues/8411 for details.
