+++ gradle.properties
@@ -19,6 +19,7 @@ kotlinVersion = 1.5.21
 kotlinCoroutinesVersion = 1.5.0
 kotlinxSerializationVersion = 1.2.1
 
+androidPluginVersion = 4.2.2
 classGraphVersion = 4.8.110
 federationGraphQLVersion = 0.6.4
 graphQLJavaVersion = 16.2
