+++ plugins/graphql-kotlin-gradle-plugin/src/test/kotlin/com/expediagroup/graphql/plugin/gradle/GraphQLGradlePluginAndroidIT.kt
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2021 Expedia, Inc
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.expediagroup.graphql.plugin.gradle
 
 import com.expediagroup.graphql.plugin.gradle.tasks.GENERATE_CLIENT_TASK_NAME
@@ -11,8 +27,7 @@ import java.io.File
 import kotlin.test.assertEquals
 
 class GraphQLGradlePluginAndroidIT {
-
-//    @Suppress("UnstableApiUsage")
+    
     @ParameterizedTest
     @MethodSource("pluginTests")
     @EnabledIfEnvironmentVariable(named = "ANDROID_SDK_ROOT", matches = ".+")
